// next-ui/src/generated/openapi/types.gen.ts
export type SearchConditionAllOfBook = {
    allOf: Array<SearchConditionBook>;
};

export type SearchConditionAllOfSeries = {
    allOf: Array<SearchConditionSeries>;
};

export type SearchConditionAnyOfBook = {
    anyOf: Array<SearchConditionBook>;
};

export type SearchConditionAnyOfSeries = {
    anyOf: Array<SearchConditionSeries>;
};

export type SearchConditionAuthor = {
    author: SearchOperatorEqualityAuthorMatch;
};

export type SearchConditionBook = SearchConditionAnyOfBook | SearchConditionAllOfBook | SearchConditionLibraryId | SearchConditionReadListId | SearchConditionSeriesId | SearchConditionDeleted | SearchConditionOneShot | SearchConditionTitle | SearchConditionReleaseDate | SearchConditionTag | SearchConditionNumberSort | SearchConditionReadStatus | SearchConditionMediaStatus | SearchConditionMediaProfile | SearchConditionAuthor | SearchConditionPoster;

export type SearchConditionCollectionId = {
    collectionId: SearchOperatorEqualityString;
};

export type SearchConditionComplete = {
    complete: SearchOperatorBoolean;
};

export type SearchConditionDeleted = {
    deleted: SearchOperatorBoolean;
};

export type SearchConditionGenre = {
    genre: SearchOperatorEqualityNullableString;
};

export type SearchConditionLanguage = {
    language: SearchOperatorEqualityString;
};

export type SearchConditionLibraryId = {
    libraryId: SearchOperatorEqualityString;
};

export type SearchConditionMediaProfile = {
    mediaProfile: SearchOperatorEqualityMediaProfile;
};

export type SearchConditionMediaStatus = {
    mediaStatus: SearchOperatorEqualityStatus;
};

export type SearchConditionNumberSort = {
    numberSort: SearchOperatorNumericTFloat;
};

export type SearchConditionOneShot = {
    oneShot: SearchOperatorBoolean;
};

export type SearchConditionPoster = {
    poster: SearchOperatorEqualityPosterMatch;
};

export type SearchConditionPublisher = {
    publisher: SearchOperatorEqualityString;
};

export type SearchConditionReadListId = {
    readListId: SearchOperatorEqualityString;
};

export type SearchConditionReadStatus = {
    readStatus: SearchOperatorEqualityReadStatus;
};

export type SearchConditionReleaseDate = {
    releaseDate: SearchOperatorDate;
};
// ...
export type SearchConditionSeriesId = {
    seriesId: SearchOperatorEqualityString;
};

export type SearchConditionSeriesStatus = {
    seriesStatus: SearchOperatorEqualityStatus;
};

export type SearchConditionSharingLabel = {
    sharingLabel: SearchOperatorEqualityNullableString;
};

export type SearchConditionTag = {
    tag: SearchOperatorEqualityNullableString;
};

export type SearchConditionTitle = {
    title: SearchOperatorString;
};

export type SearchConditionTitleSort = {
    titleSort: SearchOperatorString;
};

export type SearchOperatorAfter = {
    dateTime: Date;
};

export type SearchOperatorBefore = {
    dateTime: Date;
};

export type SearchOperatorBeginsWith = {
    value: string;
};

export type SearchOperatorBoolean = (SearchOperatorIsTrue | SearchOperatorIsFalse) & {
    operator: string;
};

export type SearchOperatorContains = {
    value: string;
};

export type SearchOperatorDate = (SearchOperatorBefore | SearchOperatorAfter | SearchOperatorIsInTheLast | SearchOperatorIsNotInTheLast | SearchOperatorIsNull | SearchOperatorIsNotNull) & {
    operator: string;
};

export type SearchOperatorDoesNotBeginWith = {
    value: string;
};

export type SearchOperatorDoesNotContain = {
    value: string;
};

export type SearchOperatorDoesNotEndWith = {
    value: string;
};

export type SearchOperatorEndsWith = {
    value: string;
};

export type SearchOperatorEqualityAuthorMatch = (SearchOperatorIs | SearchOperatorIsNot) & {
    operator: string;
};

export type SearchOperatorEqualityMediaProfile = (SearchOperatorIs | SearchOperatorIsNot) & {
    operator: string;
};

export type SearchOperatorEqualityNullableString = (SearchOperatorIs | SearchOperatorIsNot | SearchOperatorIsNullT | SearchOperatorIsNotNullT) & {
    operator: string;
};

export type SearchOperatorEqualityPosterMatch = (SearchOperatorIs | SearchOperatorIsNot) & {
    operator: string;
};

export type SearchOperatorEqualityReadStatus = (SearchOperatorIs | SearchOperatorIsNot) & {
    operator: string;
};

export type SearchOperatorEqualityStatus = (SearchOperatorIs | SearchOperatorIsNot) & {
    operator: string;
};

export type SearchOperatorEqualityString = (SearchOperatorIs | SearchOperatorIsNot) & {
    operator: string;
};
// ...
export type SearchOperatorNumericTFloat = (SearchOperatorGreaterThan | SearchOperatorLessThan | SearchOperatorIs | SearchOperatorIsNot) & {
    operator: string;
};

export type SearchOperatorString = (SearchOperatorBeginsWith | SearchOperatorDoesNotBeginWith | SearchOperatorContains | SearchOperatorDoesNotContain | SearchOperatorEndsWith | SearchOperatorDoesNotEndWith | SearchOperatorIs | SearchOperatorIsNot) & {
    operator: string;
};